refactory: 차트 fetching을 hook으로 분리한다.
- 정적인 api를 `useChartData`에서 관리한다.
- 동적인 ws를 'useChartData`에서 관리한다.
- 싱글톤 패턴을 `ChartSocketManager`에서 관리한다.
