+++ client/src/hook/index.ts
@@ -1,3 +1,4 @@
 export { default as useTrackingPage } from './global/goolgeAnalytics';
 export { default as useApiData } from './global/api/useApiData';
 export { default as useOrderBook } from './global/api/useOrderBook';
+export { default as useChartData } from './global/api/useChartData';
